add tests for collecting arbitrary artifacts
